create payment

+++ Spring_Web_MVC_Final/src/main/java/lk/ijse/spring/dto/PaymentDTO.java
@@ -16,7 +16,7 @@ public class PaymentDTO {
     private String paymentId;
     private LocalDate date;
     private double rent_amount;
-    private long extra_mileage;
+    private double extra_mileage;
     private double total;
     private double damage_cost;
     private String damageDescription;
